chore(astronvim): move 'icons.lua' to 'extras/icons.lua'
  ## what
  - Move dotfiles/astronvim/icons.lua to dotfiles/astronvim/extras/icons.lua

  ## how

  ## why
  - To keep Astronvim config folder clean
  - extra/ folder can be used for adding functionality thats not directly supported by Astronvim

  ## where

  ## usage
