+++ tools/astronvim/extras/icons.lua
